requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
groq
python-dotenv
